2014-08-20 22:57 UTC+0200 Viktor Szakats (vszakats users.noreply.github.com)   * contrib/hbwin/*.c   * contrib/hbwin/hbwapi.h     ! WAPI_IMAGELIST_GETICONSIZE(), WAPI_GETSCROLLBARINFO(),       WAPI_GETSCROLLINFO(), WAPI_GETSCROLLRANGE()       fixed to clear ref params on error     ! WAPI_TEXTOUT() fixed to return the right type on error     ! WAPI_IMAGELIST_LOADIMAGE() fixed for UNICODE     ! use NINT version of hb_ret/hb_stor/etc functions when       returning DWORD/UINT value, to keep the sign intact when       reading it on .prg level     ! use hb_retnint() functions for larger than DWORD       values (f.e. OLE error number and some more)     ! hbwapi_par_UINT() fixed to use hb_parnl() (was hb_parni())     * cleaned casts around MAKEINTRESOURCE()     * added few missing casts, deleted few unnecessary ones     * minor cleanups
  * extras/gtwvw/*.c
    * use more HB_PARHANDLE()
    * cleaned casts around MAKEINTRESOURCE()

  * contrib/hbgt/*.c
    * comment cleanup

+++ contrib/hbwin/wapi_winbase_mutex.c
@@ -100,7 +100,7 @@ HB_FUNC( WAPI_OPENMUTEX )
 {
 #if ! defined( HB_OS_WIN_CE )
    void * hName;
-   HANDLE hMutex = OpenMutex( hb_parnl( 1 ), hb_parl( 2 ), HB_PARSTR( 3, &hName, NULL ) );
+   HANDLE hMutex = OpenMutex( ( DWORD ) hb_parnl( 1 ), hb_parl( 2 ), HB_PARSTR( 3, &hName, NULL ) );
 
    hbwapi_SetLastError( GetLastError() );
    hbwapi_mutex_ret( hMutex );
